LCD functionality nearly (3/4) complete, made slight corrections in controller.

+++ controller/app/controlLCD.h
@@ -8,7 +8,7 @@
 void send_Latest_Input(char last_input);
 void send_Pattern_Name(int Pattern);
 void send_Blinking_toggle(int Toggle);
-void send_Pattern_Speed(int new_speed);
+void send_Pattern_Speed(long int new_speed);
 void init_LCD_I2C(void); 
 
 #endif // CONTROL_LCD_H
